+++ src/Symfony/Component/HttpKernel/Kernel.php
@@ -622,7 +622,7 @@ abstract class Kernel implements KernelInterface, RebootableInterface, Terminabl
      */
     protected function buildContainer(): ContainerBuilder
     {
-        foreach (['cache' => $this->getCacheDir(), 'build' => $this->warmupDir ?: $this->getBuildDir(), 'logs' => $this->getLogDir()] as $name => $dir) {
+        foreach (['cache' => $this->getCacheDir(), 'build' => $this->warmupDir ?: $this->getBuildDir()] as $name => $dir) {
             if (!is_dir($dir)) {
                 if (false === @mkdir($dir, 0777, true) && !is_dir($dir)) {
                     throw new \RuntimeException(sprintf('Unable to create the "%s" directory (%s).', $name, $dir));
@@ -630,6 +630,9 @@ abstract class Kernel implements KernelInterface, RebootableInterface, Terminabl
             } elseif (!is_writable($dir)) {
                 throw new \RuntimeException(sprintf('Unable to write in the "%s" directory (%s).', $name, $dir));
             }
+        }
+        if (!is_dir($dir = $this->getLogDir())) {
+            @mkdir($dir, 0777, true);
         }
 
         $container = $this->getContainerBuilder();